requirements.txt
six
pymongo[srv]
certifi
DateTime
from-root
pandas
numpy
scikit-learn
pyYAML
evidently
imblearn
Python-dateutil
fastapi
uvicorn
-e .